requirements.txt
# Core ML dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0

# System monitoring and utilities
psutil>=5.8.0

# Data handling and processing
Pillow>=8.3.0

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Optional: For enhanced functionality
# tensorboard>=2.8.0        # For logging (optional)
# matplotlib>=3.5.0         # For plotting (optional)
# seaborn>=0.11.0          # For enhanced plots (optional)

# Code quality (optional)
# black>=22.0.0            # Code formatting
# flake8>=4.0.0            # Linting
# mypy>=0.950              # Type checking